% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{deformula}
\alias{deformula}
\alias{deformula-package}
\title{deformula: Integration of One-Dimensional Functions with Double Exponential Formulas}
\description{
Numerical quadrature of functions of one variable over a finite
or infinite interval with double exponential formulas.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/okamumu/deformula/}
  \item Report bugs at \url{https://github.com/okamumu/deformula/issues}
}

}
\author{
\strong{Maintainer}: Hiroyuki Okamura \email{okamu@hiroshima-u.ac.jp} (\href{https://orcid.org/0000-0001-6881-0593}{ORCID})

}
