\name{deformula-package}
\alias{deformula-package}
\alias{deformula}
\docType{package}
\title{Integration of one-dimensional functions with double exponential formulas}
\description{Numerical quadrature of functions of one variable over a finite or 
infinite interval with double exponential formulas.}
\details{
\tabular{ll}{
Package: \tab deformula\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2015-10-12\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Hiroyuki Okamura

Maintainer: Hiroyuki Okamura <okamu@rel.hiroshima-u.ac.jp>
}
% \references{
% ~~ Literature or other references for background information ~~
% }
\keyword{ package }
\seealso{
\code{\link{deformula.zeroinf}}
\code{\link{deformula.moneone}}
}
